Fixed warning CS0219 at EngineerTools.cs:50,20
The variable 'decimalMaskChar' is assigned but its value is never used

The warning was related to an improper variable name in duplicate code
within the number-to-string function SimpleFormatter method use to display
human-readable values. The variable in question triggers the printing of
digits after the decimal mark and could be modified to change the number
of digits to be shown. The two behaviours seem to have been separated
at one point and the trigger was deactivated. Comments have been added
in the relevant code section.